+++ unicorn_binance_websocket_api/unicorn_binance_websocket_api_connection.py
@@ -47,29 +47,32 @@ connect = websockets.connect
 
 
 class BinanceWebSocketApiConnection(object):
-    def __init__(self, handler_binance_websocket_api_manager, stream_id, channels, markets, symbols):
-        self.handler_binance_websocket_api_manager = handler_binance_websocket_api_manager
-        self.api_key = copy.deepcopy(self.handler_binance_websocket_api_manager.stream_list[stream_id]['api_key'])
-        self.api_secret = copy.deepcopy(self.handler_binance_websocket_api_manager.stream_list[stream_id]['api_secret'])
+    def __init__(self, manager, stream_id, channels, markets, symbols):
+        self.manager = manager
+        self.api_key = copy.deepcopy(self.manager.stream_list[stream_id]['api_key'])
+        self.api_secret = copy.deepcopy(self.manager.stream_list[stream_id]['api_secret'])
         self.channels = copy.deepcopy(channels)
         self.markets = copy.deepcopy(markets)
         self.stream_id = copy.deepcopy(stream_id)
         self.symbols = copy.deepcopy(symbols)
 
     async def __aenter__(self):
-        if self.handler_binance_websocket_api_manager.is_stop_request(self.stream_id):
-            self.handler_binance_websocket_api_manager.stream_is_stopping(self.stream_id)
+        if self.manager.is_stop_request(self.stream_id):
+            self.manager.stream_is_stopping(self.stream_id)
             sys.exit(0)
-        uri = self.handler_binance_websocket_api_manager.create_websocket_uri(self.channels, self.markets,
-                                                                              self.stream_id, self.api_key,
-                                                                              self.api_secret, symbols=self.symbols)
+        uri = self.manager.create_websocket_uri(self.channels,
+                                                self.markets,
+                                                self.stream_id,
+                                                self.api_key,
+                                                self.api_secret,
+                                                symbols=self.symbols)
         if uri is False:
             # cant get a valid URI, so this stream has to crash
             error_msg = "Probably no internet connection?"
-            logging.critical("BinanceWebSocketApiConnection->await._conn.__aenter__(" + str(self.stream_id) + ", " +
+            logging.critical("BinanceWebSocketApiConnection()->await._conn.__aenter__(" + str(self.stream_id) + ", " +
                              str(self.channels) + ", " + str(self.markets) + ") - " + " error: 5 - " + str(error_msg))
-            self.handler_binance_websocket_api_manager.stream_is_crashing(self.stream_id, str(error_msg))
-            self.handler_binance_websocket_api_manager.set_restart_request(self.stream_id)
+            self.manager.stream_is_crashing(self.stream_id, str(error_msg))
+            self.manager.set_restart_request(self.stream_id)
             sys.exit(1)
         try:
             if isinstance(uri, dict):
@@ -85,136 +88,136 @@ class BinanceWebSocketApiConnection(object):
                     # -2015 = Invalid API-key, IP, or permissions for action
                     # -11001 = Isolated margin account does not exist.
                     # Cant get a valid listen_key, so this stream has to crash:
-                    logging.critical("BinanceWebSocketApiConnection->await._conn.__aenter__(" + str(self.stream_id) +
+                    logging.critical("BinanceWebSocketApiConnection()->await._conn.__aenter__(" + str(self.stream_id) +
                                      ", " + str(self.channels) + ", " + str(self.markets) + ") - " + " error: 4 - " +
                                      str(uri['msg']))
                     try:
-                        del self.handler_binance_websocket_api_manager.restart_requests[self.stream_id]
+                        del self.manager.restart_requests[self.stream_id]
                     except KeyError as error_msg:
-                        logging.critical("BinanceWebSocketApiConnection->await._conn.__aenter__(" + str(self.stream_id) +
+                        logging.critical("BinanceWebSocketApiConnection()->await._conn.__aenter__(" + str(self.stream_id) +
                                          ", " + str(self.channels) + ", " + str(self.markets) + ")" + " error: 6 - "
                                          + str(error_msg))
                     except TypeError as error_msg:
-                        logging.critical("BinanceWebSocketApiConnection->await._conn.__aenter__(" + str(self.stream_id) +
+                        logging.critical("BinanceWebSocketApiConnection()->await._conn.__aenter__(" + str(self.stream_id) +
                                          ", " + str(self.channels) + ", " + str(self.markets) + ")" + " error: 3 - "
                                          + str(error_msg))
                 else:
-                    logging.critical("BinanceWebSocketApiConnection->await._conn.__aenter__(" + str(self.stream_id) +
+                    logging.critical("BinanceWebSocketApiConnection()->await._conn.__aenter__(" + str(self.stream_id) +
                                      ", " + str(self.channels) + ", " + str(self.markets) + ") - " + " Received unknown"
                                      " error msg from Binance: " + str(uri['msg']))
-                self.handler_binance_websocket_api_manager.stream_is_crashing(self.stream_id, str(uri['msg']))
-                if self.handler_binance_websocket_api_manager.throw_exception_if_unrepairable:
+                self.manager.stream_is_crashing(self.stream_id, str(uri['msg']))
+                if self.manager.throw_exception_if_unrepairable:
                     raise StreamRecoveryError("stream_id " + str(self.stream_id) + ": " + str(uri))
                 sys.exit(1)
-        except KeyError:
-            logging.critical("BinanceWebSocketApiConnection->await._conn.__aenter__(" + str(self.stream_id) +
+        except KeyError as error_msg:
+            logging.critical("BinanceWebSocketApiConnection()->await._conn.__aenter__(" + str(self.stream_id) +
                              ", " + str(self.channels) + ", " + str(self.markets) + ")" + " error: 1 - "
                              + str(error_msg))
         self._conn = connect(uri, ping_interval=20, close_timeout=10,
-                             extra_headers={'User-Agent': 'oliver-zehentleitner/unicorn-binance-websocket-api/' +
-                                            self.handler_binance_websocket_api_manager.version})
+                             extra_headers={'User-Agent': 'unicorn-binance-websocket-api_' +
+                                            self.manager.version})
         try:
             try:
-                self.handler_binance_websocket_api_manager.websocket_list[self.stream_id] = await self._conn.__aenter__()
+                self.manager.websocket_list[self.stream_id] = await self._conn.__aenter__()
             except websockets.exceptions.InvalidMessage as error_msg:
-                logging.error("BinanceWebSocketApiConnection->await._conn.__aenter__(" + str(self.stream_id) +
+                logging.error("BinanceWebSocketApiConnection()->await._conn.__aenter__(" + str(self.stream_id) +
                               ", " + str(self.channels) + ", " + str(self.markets) + ") InvalidMessage " +
                               str(error_msg))
-                self.handler_binance_websocket_api_manager.stream_is_crashing(self.stream_id, str(error_msg))
+                self.manager.stream_is_crashing(self.stream_id, str(error_msg))
                 time.sleep(2)
-                self.handler_binance_websocket_api_manager.set_restart_request(self.stream_id)
+                self.manager.set_restart_request(self.stream_id)
                 sys.exit(1)
             except websockets.exceptions.InvalidStatusCode as error_msg:
                 if "HTTP 429" in str(error_msg):
-                    logging.error("BinanceWebSocketApiConnection->await._conn.__aenter__(" + str(self.stream_id) +
+                    logging.error("BinanceWebSocketApiConnection()->await._conn.__aenter__(" + str(self.stream_id) +
                                   ", " + str(self.channels) + ", " + str(self.markets) + ") InvalidStatusCode-HTTP429" +
                                   str(error_msg))
-                    self.handler_binance_websocket_api_manager.stream_is_crashing(self.stream_id, str(error_msg))
+                    self.manager.stream_is_crashing(self.stream_id, str(error_msg))
                     time.sleep(2)
-                    self.handler_binance_websocket_api_manager.set_restart_request(self.stream_id)
+                    self.manager.set_restart_request(self.stream_id)
                     sys.exit(1)
                 else:
-                    logging.error("BinanceWebSocketApiConnection->await._conn.__aenter__(" + str(self.stream_id) +
+                    logging.error("BinanceWebSocketApiConnection()->await._conn.__aenter__(" + str(self.stream_id) +
                                   ", " + str(self.channels) + ", " + str(self.markets) + ") InvalidStatusCode" +
                                   str(error_msg))
-            while self.handler_binance_websocket_api_manager.stream_list[self.stream_id]['payload']:
-                payload = self.handler_binance_websocket_api_manager.stream_list[self.stream_id]['payload'].pop(0)
+            while self.manager.stream_list[self.stream_id]['payload']:
+                payload = self.manager.stream_list[self.stream_id]['payload'].pop(0)
                 await self.send(json.dumps(payload, ensure_ascii=False))
-            self.handler_binance_websocket_api_manager.stream_list[self.stream_id]['status'] = "running"
-            self.handler_binance_websocket_api_manager.stream_list[self.stream_id]['has_stopped'] = False
+            self.manager.stream_list[self.stream_id]['status'] = "running"
+            self.manager.stream_list[self.stream_id]['has_stopped'] = False
             try:
-                if self.handler_binance_websocket_api_manager.restart_requests[self.stream_id]['status'] == "restarted":
-                    self.handler_binance_websocket_api_manager.increase_reconnect_counter(self.stream_id)
-                    del self.handler_binance_websocket_api_manager.restart_requests[self.stream_id]
+                if self.manager.restart_requests[self.stream_id]['status'] == "restarted":
+                    self.manager.increase_reconnect_counter(self.stream_id)
+                    del self.manager.restart_requests[self.stream_id]
             except KeyError:
                 pass
-            self.handler_binance_websocket_api_manager.set_heartbeat(self.stream_id)
+            self.manager.set_heartbeat(self.stream_id)
         except ConnectionResetError as error_msg:
-            logging.error("BinanceWebSocketApiConnection->await._conn.__aenter__(" + str(self.stream_id) + ", " +
+            logging.error("BinanceWebSocketApiConnection()->await._conn.__aenter__(" + str(self.stream_id) + ", " +
                           str(self.channels) + ", " + str(self.markets) + ")" + " - ConnectionResetError - " +
                           str(error_msg))
         except socket.gaierror as error_msg:
-            logging.critical("BinanceWebSocketApiConnection->await._conn.__aenter__(" + str(self.stream_id) + ", " +
+            logging.critical("BinanceWebSocketApiConnection()->await._conn.__aenter__(" + str(self.stream_id) + ", " +
                              str(self.channels) + ", " + str(self.markets) + ")" + " - No internet connection? "
                              "- " + str(error_msg))
-            self.handler_binance_websocket_api_manager.stream_is_crashing(self.stream_id, (str(error_msg) +
+            self.manager.stream_is_crashing(self.stream_id, (str(error_msg) +
                                                                           " - No internet connection?"))
-            self.handler_binance_websocket_api_manager.set_restart_request(self.stream_id)
+            self.manager.set_restart_request(self.stream_id)
             sys.exit(1)
         except OSError as error_msg:
-            logging.critical("BinanceWebSocketApiConnection->await._conn.__aenter__(" + str(self.stream_id) + ", " +
+            logging.critical("BinanceWebSocketApiConnection()->await._conn.__aenter__(" + str(self.stream_id) + ", " +
                              str(self.channels) + ", " + str(self.markets) + ")" + " - OSError - " + str(error_msg))
-            self.handler_binance_websocket_api_manager.stream_is_crashing(self.stream_id, (str(error_msg)))
-            self.handler_binance_websocket_api_manager.set_restart_request(self.stream_id)
+            self.manager.stream_is_crashing(self.stream_id, (str(error_msg)))
+            self.manager.set_restart_request(self.stream_id)
             sys.exit(1)
         except websockets.exceptions.InvalidStatusCode as error_msg:
             if "Status code not 101: 414" in str(error_msg):
                 # Since we subscribe via websocket.send() and not with URI anymore, this is obsolete code I guess.
-                self.handler_binance_websocket_api_manager.stream_is_crashing(self.stream_id, str(error_msg) +
+                self.manager.stream_is_crashing(self.stream_id, str(error_msg) +
                                                                               " --> URI too long?")
-                logging.critical("BinanceWebSocketApiConnection->await._conn.__aenter__(" + str(self.stream_id) + ", " +
+                logging.critical("BinanceWebSocketApiConnection()->await._conn.__aenter__(" + str(self.stream_id) + ", " +
                                  str(self.channels) + ", " + str(self.markets) + ")" + " - URI Too Long? - "
                                  + str(error_msg))
                 try:
-                    self.handler_binance_websocket_api_manager.websocket_list[self.stream_id].close()
+                    self.manager.websocket_list[self.stream_id].close()
                 except KeyError:
                     pass
                 sys.exit(1)
             elif "Status code not 101: 400" in str(error_msg):
-                logging.critical("BinanceWebSocketApiConnection->await._conn.__aenter__(" + str(self.stream_id) + ", " +
+                logging.critical("BinanceWebSocketApiConnection()->await._conn.__aenter__(" + str(self.stream_id) + ", " +
                                  str(self.channels) + ", " + str(self.markets) + ") " + str(error_msg))
             elif "Status code not 101: 429" in str(error_msg):
-                logging.error("BinanceWebSocketApiConnection->await._conn.__aenter__(" + str(self.stream_id) + ", " +
+                logging.error("BinanceWebSocketApiConnection()->await._conn.__aenter__(" + str(self.stream_id) + ", " +
                               str(self.channels) + ", " + str(self.markets) + ") " + str(error_msg))
-                self.handler_binance_websocket_api_manager.stream_is_crashing(self.stream_id, str(error_msg))
+                self.manager.stream_is_crashing(self.stream_id, str(error_msg))
                 time.sleep(2)
-                self.handler_binance_websocket_api_manager.set_restart_request(self.stream_id)
+                self.manager.set_restart_request(self.stream_id)
                 sys.exit(1)
             elif "Status code not 101: 500" in str(error_msg):
-                logging.critical("BinanceWebSocketApiConnection->await._conn.__aenter__(" + str(self.stream_id) + ", " +
+                logging.critical("BinanceWebSocketApiConnection()->await._conn.__aenter__(" + str(self.stream_id) + ", " +
                                  str(self.channels) + ", " + str(self.markets) + ") " + str(error_msg))
-                self.handler_binance_websocket_api_manager.stream_is_crashing(self.stream_id, str(error_msg))
+                self.manager.stream_is_crashing(self.stream_id, str(error_msg))
                 sys.exit(1)
             else:
-                logging.error("BinanceWebSocketApiConnection->await._conn.__aenter__(" + str(self.stream_id) + ", " +
+                logging.error("BinanceWebSocketApiConnection()->await._conn.__aenter__(" + str(self.stream_id) + ", " +
                               str(self.channels) + ", " + str(self.markets) + ") " + str(error_msg))
                 try:
-                    self.handler_binance_websocket_api_manager.websocket_list[self.stream_id].close()
+                    self.manager.websocket_list[self.stream_id].close()
                 except KeyError:
                     pass
-                self.handler_binance_websocket_api_manager.stream_is_crashing(self.stream_id, str(error_msg))
-                self.handler_binance_websocket_api_manager.set_restart_request(self.stream_id)
+                self.manager.stream_is_crashing(self.stream_id, str(error_msg))
+                self.manager.set_restart_request(self.stream_id)
                 sys.exit(1)
         except websockets.exceptions.ConnectionClosed as error_msg:
-            logging.error("BinanceWebSocketApiSocket->await._conn.__aenter__(" + str(self.stream_id) + ", " +
+            logging.error("BinanceWebSocketApiConnection()->await._conn.__aenter__(" + str(self.stream_id) + ", " +
                           str(self.channels) + ", " + str(self.markets) + ") Exception ConnectionClosed "
                           "Info: " + str(error_msg))
             if "WebSocket connection is closed: code = 1006" in str(error_msg):
-                self.handler_binance_websocket_api_manager.websocket_list[self.stream_id].close()
-                self.handler_binance_websocket_api_manager.stream_is_crashing(self.stream_id, str(error_msg))
+                self.manager.websocket_list[self.stream_id].close()
+                self.manager.stream_is_crashing(self.stream_id, str(error_msg))
                 sys.exit(1)
             else:
-                logging.critical("BinanceWebSocketApiConnection->await._conn.__aenter__(" + str(self.stream_id) +
+                logging.critical("BinanceWebSocketApiConnection()->await._conn.__aenter__(" + str(self.stream_id) +
                                  ", " + str(self.channels) + ", " + str(self.markets) + ") UnhandledException "
                                  "ConnectionClosed" + str(error_msg))
         return self
@@ -223,93 +226,93 @@ class BinanceWebSocketApiConnection(object):
         try:
             await self._conn.__aexit__(*args, **kwargs)
         except AttributeError as error_msg:
-            logging.error("binance_websocket_api_connection->__aexit__(*args, **kwargs): "
+            logging.error("BinanceWebSocketApiConnection()->__aexit__(*args, **kwargs): "
                           "AttributeError - " + str(error_msg))
         except websockets.exceptions.ConnectionClosed as error_msg:
-            logging.error("binance_websocket_api_connection->__aexit__(*args, **kwargs): "
+            logging.error("BinanceWebSocketApiConnection()->__aexit__(*args, **kwargs): "
                           "ConnectionClosed - " + str(error_msg))
-            self.handler_binance_websocket_api_manager.stream_is_stopping(self.stream_id)
-            if self.handler_binance_websocket_api_manager.is_stop_request(self.stream_id) is False and \
-                    self.handler_binance_websocket_api_manager.is_stop_as_crash_request is False:
-                self.handler_binance_websocket_api_manager.set_restart_request(self.stream_id)
+            self.manager.stream_is_stopping(self.stream_id)
+            if self.manager.is_stop_request(self.stream_id) is False and \
+                    self.manager.is_stop_as_crash_request is False:
+                self.manager.set_restart_request(self.stream_id)
             sys.exit(1)
 
     async def close(self):
-        if self.handler_binance_websocket_api_manager.is_stop_as_crash_request(self.stream_id) is False:
-            self.handler_binance_websocket_api_manager.stream_is_stopping(self.stream_id)
-        logging.info(f"binance_websocket_api_connection->close({str(self.stream_id)})")
+        if self.manager.is_stop_as_crash_request(self.stream_id) is False:
+            self.manager.stream_is_stopping(self.stream_id)
+        logging.info(f"BinanceWebSocketApiConnection()->close({str(self.stream_id)})")
         try:
-            await self.handler_binance_websocket_api_manager.websocket_list[self.stream_id].close()
+            await self.manager.websocket_list[self.stream_id].close()
         except KeyError:
-            logging.error(f"binance_websocket_api_connection->close({str(self.stream_id)}) - Stream not found!")
+            logging.error(f"BinanceWebSocketApiConnection()->close({str(self.stream_id)}) - Stream not found!")
         except RuntimeError as error_msg:
-            logging.error(f"binance_websocket_api_connection->close({str(self.stream_id)}) - "
+            logging.error(f"BinanceWebSocketApiConnection()->close({str(self.stream_id)}) - "
                           f"RuntimeError: {str(error_msg)}")
 
     async def receive(self):
-        self.handler_binance_websocket_api_manager.set_heartbeat(self.stream_id)
+        self.manager.set_heartbeat(self.stream_id)
         try:
-            received_data_json = await self.handler_binance_websocket_api_manager.websocket_list[self.stream_id].recv()
+            received_data_json = await self.manager.websocket_list[self.stream_id].recv()
             try:
-                if self.handler_binance_websocket_api_manager.restart_requests[self.stream_id]['status'] == "restarted":
-                    self.handler_binance_websocket_api_manager.increase_reconnect_counter(self.stream_id)
-                    del self.handler_binance_websocket_api_manager.restart_requests[self.stream_id]
+                if self.manager.restart_requests[self.stream_id]['status'] == "restarted":
+                    self.manager.increase_reconnect_counter(self.stream_id)
+                    del self.manager.restart_requests[self.stream_id]
             except KeyError:
                 pass
             if received_data_json is not None:
                 size = sys.getsizeof(received_data_json)
-                self.handler_binance_websocket_api_manager.increase_processed_receives_statistic(self.stream_id)
-                self.handler_binance_websocket_api_manager.add_total_received_bytes(size)
-                self.handler_binance_websocket_api_manager.increase_received_bytes_per_second(self.stream_id,
+                self.manager.increase_processed_receives_statistic(self.stream_id)
+                self.manager.add_total_received_bytes(size)
+                self.manager.increase_received_bytes_per_second(self.stream_id,
                                                                                               size)
             return received_data_json
         except RuntimeError as error_msg:
-            logging.error("binance_websocket_api_connection->receive(" +
+            logging.error("BinanceWebSocketApiConnection()->receive(" +
                           str(self.stream_id) + ") - RuntimeError - error_msg: " + str(error_msg))
-            self.handler_binance_websocket_api_manager.stream_is_stopping(self.stream_id)
-            if self.handler_binance_websocket_api_manager.is_stop_request(self.stream_id) is False:
-                self.handler_binance_websocket_api_manager.set_restart_request(self.stream_id)
+            self.manager.stream_is_stopping(self.stream_id)
+            if self.manager.is_stop_request(self.stream_id) is False:
+                self.manager.set_restart_request(self.stream_id)
             sys.exit(1)
         except ssl.SSLError as error_msg:
-            logging.error("binance_websocket_api_connection->receive(" +
+            logging.error("BinanceWebSocketApiConnection()->receive(" +
                           str(self.stream_id) + ") - ssl.SSLError - error_msg: " + str(error_msg))
             # Todo: Not sure if this needs a restart ...
-            self.handler_binance_websocket_api_manager.stream_is_stopping(self.stream_id)
-            if self.handler_binance_websocket_api_manager.is_stop_request(self.stream_id) is False:
-                self.handler_binance_websocket_api_manager.set_restart_request(self.stream_id)
+            self.manager.stream_is_stopping(self.stream_id)
+            if self.manager.is_stop_request(self.stream_id) is False:
+                self.manager.set_restart_request(self.stream_id)
             sys.exit(1)
         except KeyError as error_msg:
-            logging.error("binance_websocket_api_connection->receive(" +
+            logging.error("BinanceWebSocketApiConnection()->receive(" +
                           str(self.stream_id) + ") - KeyError - error_msg: " + str(error_msg))
-            self.handler_binance_websocket_api_manager.stream_is_stopping(self.stream_id)
-            if self.handler_binance_websocket_api_manager.is_stop_request(self.stream_id) is False:
-                self.handler_binance_websocket_api_manager.set_restart_request(self.stream_id)
+            self.manager.stream_is_stopping(self.stream_id)
+            if self.manager.is_stop_request(self.stream_id) is False:
+                self.manager.set_restart_request(self.stream_id)
             sys.exit(1)
 
     async def send(self, data):
         try:
-            await self.handler_binance_websocket_api_manager.websocket_list[self.stream_id].send(data)
-            self.handler_binance_websocket_api_manager.increase_transmitted_counter(self.stream_id)
+            await self.manager.websocket_list[self.stream_id].send(data)
+            self.manager.increase_transmitted_counter(self.stream_id)
         except websockets.exceptions.ConnectionClosed as error_msg:
-            logging.error("BinanceWebSocketApiSocket->send(" + str(self.stream_id) + ", " +
+            logging.error("BinanceWebSocketApiConnection()->send(" + str(self.stream_id) + ", " +
                           str(self.channels) + ", " + str(self.markets) + ") Exception ConnectionClosed "
                           "Info: " + str(error_msg))
-            self.handler_binance_websocket_api_manager.stream_is_crashing(self.stream_id, str(error_msg))
-            self.handler_binance_websocket_api_manager.set_restart_request(self.stream_id)
+            self.manager.stream_is_crashing(self.stream_id, str(error_msg))
+            self.manager.set_restart_request(self.stream_id)
             sys.exit(1)
         except RuntimeError as error_msg:
-            logging.error("BinanceWebSocketApiSocket->send(" + str(self.stream_id) + ", " +
+            logging.error("BinanceWebSocketApiConnection()->send(" + str(self.stream_id) + ", " +
                           str(self.channels) + ", " + str(self.markets) + ") Exception RuntimeError "
                           "Info: " + str(error_msg))
-            self.handler_binance_websocket_api_manager.stream_is_crashing(self.stream_id, str(error_msg))
-            self.handler_binance_websocket_api_manager.set_restart_request(self.stream_id)
+            self.manager.stream_is_crashing(self.stream_id, str(error_msg))
+            self.manager.set_restart_request(self.stream_id)
             sys.exit(1)
         except IndexError as error_msg:
-            logging.error("BinanceWebSocketApiSocket->send(" + str(self.stream_id) + ", " +
+            logging.error("BinanceWebSocketApiConnection()->send(" + str(self.stream_id) + ", " +
                           str(self.channels) + ", " + str(self.markets) + ") Exception IndexError "
                           "Info: " + str(error_msg))
         except KeyError as error_msg:
-            logging.error("BinanceWebSocketApiSocket->send(" + str(self.stream_id) + ", " +
+            logging.error("BinanceWebSocketApiConnection()->send(" + str(self.stream_id) + ", " +
                           str(self.channels) + ", " + str(self.markets) + ") Exception KeyError "
                           "Info: " + str(error_msg))